backend/pyproject.toml
[project]
name = "en-capsule-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.5.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "geographiclib>=2.0",
    "inflection>=0.5.1",
    "pyyaml>=6.0.2",
    "uritemplate>=4.1.1",
]
